Make owner the user named same as tenant/account
This maps tenants to account in Swift. The user who's able to give
ACL / Create Containers permision will be the same user as the account.

Example: If we have the account called hellocorp with a user
hellocorp that user will be admin on that account and can give ACL
to all other users for hellocorp.

If there is a user who has not the same name the account and if it
is inside the group (or roles in keystone lingua)
keystone_admin_group as specifed in the configuration variable
(Admin by default) it will be allowed to be an admin the account.

Change-Id: Icfabea296f54d8192ad2af2fb175a27c4a65ef19

Conflicts:

	keystone/middleware/swift_auth.py